paper_researcher/requirements.txt
arxiv>=2.1.0
openai>=1.0.0
pymupdf>=1.23.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
tenacity>=8.2.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0
